change vlan id to uint16

+++ services/network/network.go
@@ -224,7 +224,7 @@ type VirtualNetworkProperties struct {
 	// Subnets that could hold ipv4 and ipv6 subnets
 	Subnets *[]Subnet `json:"subnets,omitempty"`
 	// Vlan
-	Vlan *int32 `json:"vlan,omitempty"`
+	Vlan *uint16 `json:"vlan,omitempty"`
 	// Statuses - Status
 	Statuses map[string]*string `json:"statuses"`
 }
